apps/ml/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
pymongo==4.6.1
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.1
xgboost==2.0.3
lightgbm==4.3.0
joblib==1.3.2
mlflow==2.10.1
redis==5.0.1
boto3==1.34.28
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

